+++ changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Use screen reader text for open new window hint
+- Remove filter to use unsafe URLs after they have been used
+- Updated dependencies
+
+### Fixed
+
+- Fixed a wrong placeholder to use an external link to Dropbox
+
 ## [5.3.0] - 16.08.2026
 
 ### Added